move create new addItem component, move user msg to alert windows, make shopping lists navigate to the list view
